ED-6998: always reference state from view in emoji plugin
This avoids using stale state if the editor ends up being re-created. We already
keep a reference to the EditorView, so this doesn't change the structure of the
plugin much.

Also updates the plugin's EditorView reference on an update.

The emoji refactor is Coming Soon(tm), which will simplify this plugin. Today we
just want to fix this issue so that we can release full width mode to products.