BUILDTOOLS-256 Add dependencies on all the things in multi-entry-tools
Not too happy about needing this. But the problem is that paths is always relative
to the tsconfig. So the relative paths config doesn't work for transitive dependencies.
i.e. when importing button from moduleX not using theme
inside button, theme is imported
this import will fail because the node_modules of moduleX doesn't have a copy of theme
because moduleX doesn't directly depend on it.

This "fix" I did creates a top level node_modules we can rely on to have all the right
symlinks to all the ak modules.