Fix infinite loop in bb releases addon and improve diffstat
An infinite loop bug existed in cases where a PR had over 500 commits in
the diffstat with master. This is now fixed, and the diffstat
functionality has been changed to diff against the destination branch
rather than master.

With this diffstat change, we also re-enable the plugin on merged PRs
as they will now properly show the released packages.